В контроллеры добавлены типы возвращаемых ответов, рефакторинг типов запросов в контоллерах и на фронте

+++ WebApi/Controllers/RoleController.cs
@@ -2,7 +2,9 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Services.Abstractions;
+using Services.Contracts.Group;
 using Services.Contracts.Role;
+using WebApi.Models.Group;
 using WebApi.Models.Role;
 using WebApi.Models.User;
 
@@ -24,44 +26,66 @@ namespace WebApi.Controllers
         }
 
         [HttpGet("{id}")]
+        [ProducesResponseType(typeof(RoleModel), 200)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(404)]
         public async Task<IActionResult> GetAsync(Guid id)
         {
-            return Ok(_mapper.Map<RoleModel>(await _service.GetByIdAsync(id)));
+
+            var role = await _service.GetByIdAsync(id);
+            if (role == null)
+                return NotFound();            
+            return Ok(_mapper.Map<RoleModel>(role));
         }
 
         [HttpPost]
+        [ProducesResponseType(200)]
+        [ProducesResponseType(400)]
         public async Task<IActionResult> CreateAsync(CreatingRoleModel roleModel)
         {
             return Ok(await _service.CreateAsync(_mapper.Map<CreatingRoleDto>(roleModel)));
         }
 
         [HttpPut("{id}")]
+        [ProducesResponseType(204)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(404)]
         public async Task<IActionResult> EditAsync(Guid id, UpdatingRoleModel roleModel)
         {
+            if (await _service.GetByIdAsync(id) == null)
+                return NotFound();
             await _service.UpdateAsync(id, _mapper.Map<UpdatingRoleModel, UpdatingRoleDto>(roleModel));
-            return Ok();
+            return NoContent();
         }
 
         [HttpDelete("{id}")]
+        [ProducesResponseType(204)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(404)]
         public async Task<IActionResult> DeleteAsync(Guid id)
         {
+            if (await _service.GetByIdAsync(id) == null)
+                return NotFound();
             await _service.DeleteAsync(id);
-            return Ok();
+            return NoContent();
         }
 
-        [HttpPost("list")]
+        [HttpGet("list")]
+        [ProducesResponseType(typeof(IEnumerable<RoleModel>), 200)]
         public async Task<IActionResult> GetListAsync()
         {
             return Ok(_mapper.Map<List<RoleModel>>(await _service.GetListAsync()));
         }
 
-        [HttpPost("userlist")]
+        [HttpGet("userlist")]
+        [ProducesResponseType(typeof(IEnumerable<UserModel>), 200)]
         public async Task<IActionResult> GetUserListAsync(Guid id)
         {
             return Ok(_mapper.Map<List<UserModel>>(await _service.GetUserListAsync(id)));
         }
 
-        [HttpPost("usernotinrolelist")]
+        [HttpGet("usernotinrolelist")]
+        [ProducesResponseType(typeof(IEnumerable<UserModel>), 200)]
         public async Task<IActionResult> GetUserNotInRoleListAsync(Guid id)
         {
             return Ok(_mapper.Map<List<UserModel>>(await _service.GetUserNotInRoleListAsync(id)));
