+++ WebApi/Controllers/UserController.cs
@@ -10,6 +10,8 @@ using WebApi.Models.User;
 using WebApi.Models.Group;
 using WebApi.Models.Role;
 using Microsoft.AspNetCore.Cors.Infrastructure;
+using Domain.Entities;
+using Services.Contracts.Role;
 
 namespace WebApi.Controllers
 {
@@ -29,58 +31,65 @@ namespace WebApi.Controllers
         }
 
         [HttpGet("{id}")]
+        [ProducesResponseType(typeof(UserModel), 200)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(404)]
         public async Task<IActionResult> GetAsync(Guid id)
         {
-            return Ok(_mapper.Map<UserModel>(await _service.GetByIdAsync(id)));
+            var user = await _service.GetByIdAsync(id);
+            if (user == null)
+                return NotFound();
+            return Ok(_mapper.Map<UserModel>(user));
         }
 
         [HttpPost]
+        [ProducesResponseType(200)]
+        [ProducesResponseType(400)]
         public async Task<IActionResult> CreateAsync(CreatingUserModel userModel)
         {
             return Ok(await _service.CreateAsync(_mapper.Map<CreatingUserDto>(userModel)));
         }
 
         [HttpPut("{id}")]
+        [ProducesResponseType(204)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(404)]
         public async Task<IActionResult> EditAsync(Guid id, UpdatingUserModel userModel)
         {
+            if (await _service.GetByIdAsync(id) == null)
+                return NotFound();
             await _service.UpdateAsync(id, _mapper.Map<UpdatingUserModel, UpdatingUserDto>(userModel));
-            return Ok();
+            return NoContent();
         }
 
         [HttpDelete("{id}")]
+        [ProducesResponseType(204)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(404)]
         public async Task<IActionResult> DeleteAsync(Guid id)
         {
+            if (await _service.GetByIdAsync(id) == null)
+                return NotFound();
             await _service.DeleteAsync(id);
-            return Ok();
+            return NoContent();
         }
 
-        [HttpPost("filteredlist")]
-        public async Task<IActionResult> GetPagedAsync(UserFilterModel filterModel)
-        {
-            var filterDto = _mapper.Map<UserFilterModel, UserFilterDto>(filterModel);
-            return Ok(_mapper.Map<List<UserModel>>(await _service.GetPagedAsync(filterDto)));
-        }
-
-        [HttpPost("login")]
-        public async Task<IActionResult> GetAsyncByEmailPassword(UserLoginModel userLoginModel)
-        {
-
-            return Ok(_mapper.Map<UserModel>(await _service.Login(_mapper.Map<UserLoginModel, UserLoginDto>(userLoginModel))));
-        }
-
-        [HttpPost("list")]
+        [HttpGet("list")]
+        [ProducesResponseType(typeof(IEnumerable<UserModel>), 200)]
         public async Task<IActionResult> GetListAsync()
         {            
             return Ok(_mapper.Map<List<UserModel>>(await _service.GetListAsync()));
         }
 
-        [HttpPost("grouplist")]
+        [HttpGet("grouplist")]
+        [ProducesResponseType(typeof(IEnumerable<GroupModel>), 200)]
         public async Task<IActionResult> GetGroupListAsync(Guid id)
         {
             return Ok(_mapper.Map<List<GroupModel>>(await _service.GetGroupListAsync(id)));
         }
 
-        [HttpPost("rolelist")]
+        [HttpGet("rolelist")]
+        [ProducesResponseType(typeof(IEnumerable<RoleModel>), 200)]
         public async Task<IActionResult> GetRoleListAsync(Guid id)
         {
             return Ok(_mapper.Map<List<RoleModel>>(await _service.GetRoleListAsync(id)));
